LayoutTests/platform/ios-sim-deprecated/http/tests/xmlhttprequest/XMLHttpRequestException-expected.txt

CONSOLE MESSAGE: line 1: TypeError: undefined is not an object (evaluating 'e.toString')
Tests the properties of the XMLHttpRequestException object.
